fix: navigation inside PP and ToS page
