+++ packages/pn-commons/src/hooks/index.ts
@@ -5,7 +5,6 @@ import { useIsCancelled } from './useIsCancelled';
 import { useIsMobile } from './useIsMobile';
 import { useMultiEvent } from './useMultiEvent';
 import { useProcess } from './useProcess';
-import { useRewriteLinks } from './useRewriteLinks';
 import { useSessionCheck } from './useSessionCheck';
 import { useTracking } from './useTracking';
 import { useUnload } from './useUnload';
@@ -19,7 +18,6 @@ export {
   useTracking,
   useDownloadDocument,
   useProcess,
-  useRewriteLinks,
   useHasPermissions,
   downloadDocument,
   useIsCancelled,
