+++ packages/pn-commons/src/utility/index.ts
@@ -36,7 +36,7 @@ import {
   parseNotificationDetail,
   populatePaymentsPagoPaF24,
 } from './notification.utility';
-import { compileOneTrustPath } from './onetrust.utility';
+import { compileOneTrustPath, rewriteLinks } from './onetrust.utility';
 import { calculatePages } from './pagination.utility';
 import {
   PAYMENT_CACHE_KEY,
@@ -121,4 +121,5 @@ export {
   setPaymentsInCache,
   checkIfPaymentsIsAlreadyInCache,
   PAYMENT_CACHE_KEY,
+  rewriteLinks,
 };
