+++ packages/pn-pa-webapp/src/pages/PrivacyPolicy.page.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
 
-import { compileOneTrustPath, useRewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
+import { compileOneTrustPath, rewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
 
 import * as routes from '../navigation/routes.const';
 import { getConfiguration } from '../services/configuration.service';
@@ -15,8 +15,6 @@ declare const OneTrust: {
 };
 
 const PrivacyPolicyPage = () => {
-  const [contentLoaded, setContentLoaded] = useState(false);
-
   const configuration = useMemo(() => getConfiguration(), []);
 
   const { ONE_TRUST_PP, ONE_TRUST_DRAFT_MODE } = configuration;
@@ -33,10 +31,8 @@ const PrivacyPolicyPage = () => {
   }, []);
 
   void waitForElement('.otnotice-content').then(() => {
-    setContentLoaded(true);
+    rewriteLinks(routes.PRIVACY_POLICY, '.otnotice-content a');
   });
-
-  useRewriteLinks(contentLoaded, routes.PRIVACY_POLICY, '.otnotice-content a');
 
   return (
     <>
