+++ packages/pn-pa-webapp/src/pages/TermsOfService.page.tsx
@@ -1,5 +1,6 @@
-import { useEffect, useMemo, useState } from 'react';
-import { compileOneTrustPath, useRewriteLinks } from '@pagopa-pn/pn-commons';
+import { useEffect, useMemo } from 'react';
+
+import { compileOneTrustPath, rewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
 
 import * as routes from '../navigation/routes.const';
 import { getConfiguration } from '../services/configuration.service';
@@ -14,8 +15,6 @@ declare const OneTrust: {
 };
 
 const TermsOfServicePage = () => {
-  const [contentLoaded, setContentLoaded] = useState(false);
-
   const configuration = useMemo(() => getConfiguration(), []);
 
   const { ONE_TRUST_TOS, ONE_TRUST_DRAFT_MODE } = configuration;
@@ -24,19 +23,16 @@ const TermsOfServicePage = () => {
     if (ONE_TRUST_TOS) {
       OneTrust.NoticeApi.Initialized.then(function () {
         OneTrust.NoticeApi.LoadNotices(
-          [
-            compileOneTrustPath(
-              ONE_TRUST_TOS,
-              ONE_TRUST_DRAFT_MODE
-            ),
-          ],
+          [compileOneTrustPath(ONE_TRUST_TOS, ONE_TRUST_DRAFT_MODE)],
           false
         );
-        setContentLoaded(true);
       });
     }
   }, []);
-  useRewriteLinks(contentLoaded, routes.TERMS_OF_SERVICE, '.otnotice-content a');
+
+  void waitForElement('.otnotice-content').then(() => {
+    rewriteLinks(routes.TERMS_OF_SERVICE, '.otnotice-content a');
+  });
 
   return (
     <>
