+++ packages/pn-personafisica-webapp/src/pages/PrivacyPolicy.page.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useMemo } from 'react';
 
-import { compileOneTrustPath, useRewriteLinks } from '@pagopa-pn/pn-commons';
+import { compileOneTrustPath, rewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
 
 import * as routes from '../navigation/routes.const';
 import { getConfiguration } from '../services/configuration.service';
@@ -15,8 +15,9 @@ declare const OneTrust: {
 };
 
 const PrivacyPolicyPage = () => {
-  const [contentLoaded, setContentLoaded] = useState(false);
-  const { ONE_TRUST_DRAFT_MODE, ONE_TRUST_PP } = getConfiguration();
+  const configuration = useMemo(() => getConfiguration(), []);
+
+  const { ONE_TRUST_PP, ONE_TRUST_DRAFT_MODE } = configuration;
 
   useEffect(() => {
     if (ONE_TRUST_PP) {
@@ -25,12 +26,13 @@ const PrivacyPolicyPage = () => {
           [compileOneTrustPath(ONE_TRUST_PP, ONE_TRUST_DRAFT_MODE)],
           false
         );
-        setContentLoaded(true);
       });
     }
   }, []);
 
-  useRewriteLinks(contentLoaded, routes.PRIVACY_POLICY, '.otnotice-content a');
+  void waitForElement('.otnotice-content').then(() => {
+    rewriteLinks(routes.PRIVACY_POLICY, '.otnotice-content a');
+  });
 
   return (
     <>
