+++ packages/pn-personafisica-webapp/src/pages/TermsOfService.page.tsx
@@ -1,8 +1,9 @@
-import { useEffect, useState } from 'react';
-import { compileOneTrustPath, useRewriteLinks } from '@pagopa-pn/pn-commons';
+import { useEffect, useMemo } from 'react';
+
+import { compileOneTrustPath, rewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
 
 import * as routes from '../navigation/routes.const';
-import { getConfiguration } from "../services/configuration.service";
+import { getConfiguration } from '../services/configuration.service';
 
 declare const OneTrust: {
   NoticeApi: {
@@ -14,26 +15,24 @@ declare const OneTrust: {
 };
 
 const TermsOfServicePage = () => {
-  const [contentLoaded, setContentLoaded] = useState(false);
-  const { ONE_TRUST_DRAFT_MODE, ONE_TRUST_TOS} = getConfiguration();
+  const configuration = useMemo(() => getConfiguration(), []);
+
+  const { ONE_TRUST_TOS, ONE_TRUST_DRAFT_MODE } = configuration;
 
   useEffect(() => {
     if (ONE_TRUST_TOS) {
       OneTrust.NoticeApi.Initialized.then(function () {
         OneTrust.NoticeApi.LoadNotices(
-          [
-            compileOneTrustPath(
-              ONE_TRUST_TOS,
-              ONE_TRUST_DRAFT_MODE
-            ),
-          ],
+          [compileOneTrustPath(ONE_TRUST_TOS, ONE_TRUST_DRAFT_MODE)],
           false
         );
-        setContentLoaded(true);
       });
     }
   }, []);
-  useRewriteLinks(contentLoaded, routes.TERMS_OF_SERVICE, '.otnotice-content a');
+
+  void waitForElement('.otnotice-content').then(() => {
+    rewriteLinks(routes.TERMS_OF_SERVICE, '.otnotice-content a');
+  });
 
   return (
     <>
