+++ packages/pn-personagiuridica-webapp/src/pages/PrivacyPolicy.page.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
-import { compileOneTrustPath, useRewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
+import { compileOneTrustPath, rewriteLinks, waitForElement } from '@pagopa-pn/pn-commons';
 
 import * as routes from '../navigation/routes.const';
 import { getConfiguration } from '../services/configuration.service';
@@ -16,7 +16,6 @@ declare const OneTrust: {
 
 const PrivacyPolicyPage = () => {
   const { ONE_TRUST_DRAFT_MODE, ONE_TRUST_PP } = getConfiguration();
-  const [contentLoaded, setContentLoaded] = useState(false);
 
   useEffect(() => {
     if (ONE_TRUST_PP) {
@@ -30,10 +29,8 @@ const PrivacyPolicyPage = () => {
   }, []);
 
   void waitForElement('.otnotice-content').then(() => {
-    setContentLoaded(true);
+    rewriteLinks(routes.PRIVACY_POLICY, '.otnotice-content a');
   });
-
-  useRewriteLinks(contentLoaded, routes.PRIVACY_POLICY, '.otnotice-content a');
 
   return (
     <>
